Home

SurrogateModelOptim.jl

SurrogateModelOptim._scale

SurrogateModelOptim._scale

SurrogateModelOptim._scale

SurrogateModelOptim._setup_parameters

SurrogateModelOptim._scale — Method.

function _scale(oldX::Array{T,1},newMin,newMax;oldMin=minimum(oldX), oldMax=maximum(oldX)) where T <: Real

Scale a vector to match a new range.

source

SurrogateModelOptim._scale — Method.

function _scale(oldX::Array{T,2},direction::Int,newMin,newMax; oldMin=minimum(oldX,direction)::Array{T,2},oldMax=maximum(oldX,direction)::Array{T,2}) where T <: Real

Scale a 2D matrix to match a new range along the specified direction.

source

SurrogateModelOptim._scale — Method.

function _scale(oldX::T,newMin,newMax;oldMin=minimum(oldX), oldMax=maximum(oldX)) where T <: Real

Scale a scalar to match a new range.

source

SurrogateModelOptim._setup_parameters — Function.

_setup_parameters(SearchRange, NumDimensions=false; kwargs...)

Internal function of SurrogateModelOptim to handle the interface which constrains the input combinations the user can use.

source