Home

SurrogateModelOptim.jl

SurrogateModelOptim.RBFHypersResult

SurrogateModelOptim.RMSErrorLOO

SurrogateModelOptim._rippa

SurrogateModelOptim._scale

SurrogateModelOptim._scale

SurrogateModelOptim._scale

SurrogateModelOptim._surrogate_interpolant

SurrogateModelOptim._surrogate_interpolant

SurrogateModelOptim._update_options

SurrogateModelOptim.interp_obj

SurrogateModelOptim.RBFHypersResult — Type.

RBFHypersResult(width::S,kernelFunc,scaling::U,fitness::Float64)

Datastructure to store results from the optimisation of an RBF interpolation kernel

source

SurrogateModelOptim.RMSErrorLOO — Method.

function RMSErrorLOO(interp,samples::Array{Float64,2},plan;
cond_max = 1/eps(Float64)/10000, rippa = false)

Calculate the Leave-One-Out RMS error for a interpolation method. rippa can be used to calculate the approximation of the LOO error for Radial Basis Functions at a cost of 𝛰(3) (compared to 𝛰(4)). cond_max sets the maximum allowed condition number for matrix A used in the RBF calculation.

source

SurrogateModelOptim._rippa — Method.

function _rippa(A,a)

Estimate the Leave-One-Out (LOO) errors using rippas method. Complexity of 𝛰(3) compared to calculating the exact LOO at a cost of 𝛰(4). A is the RBF matrix and a is the weights of the RBF.

source

SurrogateModelOptim._scale — Method.

function _scale(oldX::Array{T,1},newMin,newMax;oldMin=minimum(oldX), oldMax=maximum(oldX)) where T <: Real

Scale a vector to match a new range.

source

SurrogateModelOptim._scale — Method.

function _scale(oldX::Array{T,2},direction::Int,newMin,newMax; oldMin=minimum(oldX,direction)::Array{T,2},oldMax=maximum(oldX,direction)::Array{T,2}) where T <: Real

Scale a 2D matrix to match a new range along the specified direction.

source

SurrogateModelOptim._scale — Method.

function _scale(oldX::T,newMin,newMax;oldMin=minimum(oldX), oldMax=maximum(oldX)) where T <: Real

Scale a scalar to match a new range.

source

SurrogateModelOptim._surrogate_interpolant — Method.

_surrogate_interpolant(optres::T,points,observations,estimationpoints,
oldMin,oldMax) where T <: RBFoptim_v1.HypersResult{U,Float64} where U

Evaluate an optimised interpolant at locations estimationpoints.

source

SurrogateModelOptim._surrogate_interpolant — Method.

_surrogate_interpolant(optres::T,plan,samples,estimationpoints,
oldMin,oldMax) where T <: SurrogateModelOptim.RBFHypersResult{U,Float64} where U

Evaluate an optimised interpolant at locations estimationpoints.

source

SurrogateModelOptim._update_options — Method.

_update_options(search_range;kwargs...)

Internal function of SurrogateModelOptim to update the options to suitable default values.

source

SurrogateModelOptim.interp_obj — Method.

function interp_obj(inpt::Vector{Float64}, kerns, samples::Array{Float64,2},
plan::Array{Float64,2}; rippa::Bool = false, scale::Bool = false)

Objective function for optimisation of interpolation kernel function and width.

source